move list view
